extract ClientRequestWatcher to separate file
